Structured converters into directory
