+++ converters/convertToInt.js
@@ -1,0 +1,13 @@
+function convertToInt (array, options) {
+    
+    if (!options.reversed) {
+        array.reverse()
+    }
+
+    for (let i = 0; i < array.length; i++) {
+        
+    }
+
+}
+
+export default convertToInt;
