+++ script.js
@@ -1,0 +1,10 @@
+import convertToBin from "./converters/convertToBin";
+import convertToInt from "./converters/convertToInt";
+
+function intToBin (integer, size, options) {
+
+}
+
+function binToInt (array, options) {
+    
+}
